+++ apps/ledger-live-desktop/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ledger-live-desktop
 
+## 2.113.2-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`422c046`](https://github.com/LedgerHQ/ledger-live/commit/422c0468aa865a0840bb6c23506173621b0930be)]:
+  - @ledgerhq/live-common@34.33.2-hotfix.0
+  - @ledgerhq/live-dmk-desktop@0.3.6-hotfix.0
+
 ## 2.113.1
 
 ### Patch Changes